% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfi_to_osm_data_prep.R
\name{nfi_to_osm}
\alias{nfi_to_osm}
\title{NFI to OSM data}
\usage{
nfi_to_osm(
  nfi_folder,
  scenario_name = "nfi_simulation",
  remeasurement_number = NULL,
  calculate_stockable = TRUE,
  provinces = c("New Brunswick", "Prince Edward Island", "Nova Scotia",
    "Newfoundland and Labrador"),
  model_variant = c("Acadian", "Newfoundland and Labrador"),
  include_small_trees = TRUE,
  dbh_filter = 0,
  use_projected_height = TRUE,
  calculate_ef_for_dbh_bin = TRUE,
  ef_dbh_increment_bin = 4,
  output_path = getwd()
)
}
\arguments{
\item{nfi_folder}{String. The path to the folder containing the NFI data.}

\item{scenario_name}{String. The name of the OSM growth scenario to be populated into the 'Scenario' column of Stand List Table.}

\item{remeasurement_number}{Integer. NFI numbers to individualize each field campaign. Starts at '0' (plot establishment) and the following numbers in the sequence are remeasurements.}

\item{calculate_stockable}{Logical. Refer to the optional OSM column "Stockable" defaults to \code{TRUE}.
OSM uses this information to adjust calculation when projecting a stand growth. Use if it is known that there are portions of the site that are portions of the plot that do not support tree growth (e.g. roads or other).

On NFI, this information may be contained in two variables: \code{meas_plot_size} and \code{nom_plot_size}. The actual measured area (in ha) of plots (meas_plot_size) may be different than the designed size of the plot (nom_plot_size). nominal plot size is standardized to 400 square meters (or 0.04 ha).
This may happen because sites are assigned to the center of randomly established photo plots and may fall partially on non-forested areas e.g. road, or other non-productive areas.

However, it is important to note that there may be other reasons for not measuring the entire plot (fallen tree, inaccessible private property or other safety reason).

If \code{TRUE} will calculate ratio of measured size (ha) to nominal size (ha) and assume unmeasured portions are unproductive, otherwise will use nominal size (ha) and assume the entire plot is productive.
Also, if using \code{meas_num = 0} there is a higher chance fully productive plots being incompletely measured (due to cost issues). See page 12 of \link[https://nfi.nfis.org/resources/groundplot/GP_compilation_procedures_2.4.pdf]{Compilation Procedures}}

\item{provinces}{Vector of strings. Provinces for which you want data for. Will only retrieve data for one of the four provinces in the argument default.}

\item{model_variant}{Vector of string. OSM has two models. The "Acadian" model variant has been calibrated for the provinces of New Brunswick, Prince Edward Island and Nova Scotia. The "Newfoundland and Labrador" model has been calibrated for Newfoundland and Labrador only.}

\item{include_small_trees}{Logical. Defaults to true to maximize data use. Will include all individual trees measured on the small plot data. See page 32 of \link\link{https://nfi.nfis.org/resources/groundplot/Gp_guidelines_v5.0.pdf}(Ground Plot Guidelines).}

\item{dbh_filter}{Numeric. A field to filter out trees based on DBH. Will pick DBH higher than or equal to this value.}

\item{use_projected_height}{Logical. When both (measured, calculated or estimated) height and projected height are available. Prefer projected height. Projected height are calculated for broken stems (i.e. stem_cond == "B").}

\item{calculate_ef_for_dbh_bin}{Logical. Refer to the required OSM column "Stems" and defaults to \code{TRUE}.
Calculates expansion factors for DBH bins specified in \code{ef_dbh_increment_bin}, instead of for each individual tree. The calculation will first group trees into DBH class sizes (e.g. (0-4],(4-8],(8-12], and etc.).
It then aggregates data by counting number of trees and averaging DBH and height for each bin.
Finally the 'Stems' variable is calculate as the count of stem/measured_plot_size (generally 0.04 but sometimes less).

Aggregated Stems, DBH and Height will represent individual trees to be used as input in OSM.
If \code{FALSE} it will assume each individual tree has an expansion factor of 1/measured_plot_size. This may be less prefered as it gives equal weights to all individual trees in the plot.}

\item{ef_dbh_increment_bin}{Numeric. In centimeters. Only used if \code{calculate_ef_for_dbh_bin = TRUE}.}

\item{output_path}{String. Use to output the SQLite database that can be passed to OSM.}
}
\value{
Returns a list with two data frames (data.tables) and a writes a database on specified output_path. The data frames should match the input data exactly as in \link[https://forusresearch.com/downloads/osm/help/OSM.HelpFiles/OSM.Input.htm]{Open Stand Model}.
Multiple assumptions are made on the creation of each variable and they are described below.
\itemize{
\item \strong{STAND LIST TABLE}
\describe{
\item{Scenario}{User specified}
\item{SurveyID}{Same as \code{nfi_plot} in NFI data.}
\item{SurveyYear}{The year from \code{meas_date} in \emph{all_gp_site_info.csv} NFI data.}
\item{SurveyAge}{The stand age from \code{site_age} in \emph{all_gp_ltp_header.csv} NFI data. Uses arithmetic average age of trees deemed suitable for compiling site age. For more details see page 129 of \link\link{https://nfi.nfis.org/resources/groundplot/Gp_guidelines_v5.0.pdf}(Ground Plot Guidelines).}
\item{Plots}{Number of plots for a given stand. Defaults to 1 where each NFI plot will assumed to represent a stand.}
\item{Stockable}{Ratio of total sample area (all plots) that supports productive tree growth. Calculated as described in \code{calculate_stockable}}
\item{X}{NFI plots uses UTM coordinates systems which are converted to Latitude/Longitude system using package \pkg{sf}}
\item{Y}{Same as above.}
}
\item \strong{TREE LIST TABLE}
\describe{
\item{SurveyID}{Same as \code{nfi_plot} in NFI data.}
\item{TreeID}{Same as \code{tree_num} in \emph{all_gp_lgtree_plot.csv} and \code{smtree_num} in \emph{all_gp_smtree_plot.csv}}
\item{Species}{NFI Species are mapped to OSM species based on the \code{Acadian_SpeciesList} table contained in the \link[https://forusresearch.com/downloads/osm/OSMv2.25.1/OSMv2.25.1_Demo.zip]{OSM demo package}.
NFI codes are mapped to the \code{GENUS} and \code{SPECIES} column of the \code{Acadian_SpeciesList} and the variable \code{OSM_AD_CmdKey} from that table is used.}
\item{DBH}{DBH of individual trees or average DBH for each DBH Bin classes when \code{calculate_ef_for_dbh_bin = TRUE}.}
\item{HT}{Height of individual trees or average height for each DBH Bin classes when \code{calculate_ef_for_dbh_bin = TRUE}.}
\item{Stems}{Trees per hectare this record represents. i.e., tree expansion factor. Calculated as described in the argument \code{calculate_ef_for_dbh_bin}.}
\item{ToCut}{Not extract from NFI data. Defaults to NaN in OSM. See OSM input page for \link[https://forusresearch.com/downloads/osm/help/OSM.HelpFiles/OSM.Input.htm]{details}.}
\item{Weight}{Not extract from NFI data. Defaults to NaN in OSM. See OSM input page for \link[https://forusresearch.com/downloads/osm/help/OSM.HelpFiles/OSM.Input.htm]{details}.}
\item{HTK}{Not extract from NFI data. Defaults to NaN in OSM. See OSM input page for \link[https://forusresearch.com/downloads/osm/help/OSM.HelpFiles/OSM.Input.htm]{details}. Could potentially be extracted from \code{height} of \code{stem_cond = "B"} while using \code{height_prj} as OSM \code{HT}.}
\item{CR}{Crown Ratio. Calculated for both large and small trees when, at minimum, \code{crown_cond} and \code{crown_base} data are avaialble but may also use \code{crow_top} if avaialble. Calculated as \code{(crown_top-crown_base)/HT} or \code{(crown_top-HT)/HT}.}
\item{DBHI}{Not extract from NFI data. Defaults to NaN in OSM. See OSM input page for \link[https://forusresearch.com/downloads/osm/help/OSM.HelpFiles/OSM.Input.htm]{details}.}
\item{HTI}{Not extract from NFI data. Defaults to NaN in OSM. See OSM input page for \link[https://forusresearch.com/downloads/osm/help/OSM.HelpFiles/OSM.Input.htm]{details}.}
\item{Origin}{Not extract from NFI data. Defaults to NaN in OSM. See OSM input page for \link[https://forusresearch.com/downloads/osm/help/OSM.HelpFiles/OSM.Input.htm]{details}. May be partially inferred from \emph{plot_origin.csv} variables \code{veg_origin} and \code{regen_type}}
\item{Born}{Derived from \code{age_total} variable of \emph{all_gp_ltp_tree_age.csv}. For more details see psgr 27 of \link[https://nfi.nfis.org/resources/groundplot/GP_compilation_procedures_2.4.pdf]{Compilation proceedures}.}
\item{Died}{Derived from \code{lgtree_status} (Large trees) or \code{smtree_status} (Small trees). If classified as DS = Dead Standing then it is assigned 9999 to be initialized in the snag pool. LS (Live Stading), LF (Live fallen) or M (Missing) are assumed to as live (set to 0).
Setting to 9999 assumes tree died 7 years ago as per OSM default. Could potentially be improved to check whether tree died between remeasurement and set to value to be year of measurement 2 minus year of measurement 1 divided 2.)}
\item{Risk}{Not extract from NFI data. Defaults to NaN in OSM. See OSM input page for \link[https://forusresearch.com/downloads/osm/help/OSM.HelpFiles/OSM.Input.htm]{details}.}
\item{Grade}{Not extract from NFI data. Defaults to NaN in OSM. See OSM input page for \link[https://forusresearch.com/downloads/osm/help/OSM.HelpFiles/OSM.Input.htm]{details}.}
}
}
}
\description{
This function extracts and properly formats Canada's National Forest Inventory (NFI) data to Open Stand Mode (OSM) input format.
It attempts to maximize the use of data but there are still some improvements that can be made to use the full extent of NFI data. Check 'Value' section for more details.
}
\details{
It uses the current NFI data structure in csv files and is based on file naming patterns for ground plot data e.g. "all_gp_trees".
It maximizes the use of data by making several assumptions on which data should be included and how. Several of those assumptions can be changed by the user by modifying the function parameters e.g. calculate_ef_for_dbh_bin = TRUE.
For full details on how each OSM input variable is constructed and which assumptions are made consult the details.
}
